requirements.txt
Django==5.0.2
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
django-allauth==0.60.1
django-environ==0.11.2
Pillow==10.2.0
python-dotenv==1.0.1
django-ckeditor==6.7.1
django-storages==1.14.2
boto3==1.34.34
psycopg2-binary==2.9.9
gunicorn==21.2.0 